@@ -4,32 +4,31 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build-lib": "ng build ngx-modialog --prod && ng build bootstrap --prod && ng build vex --prod && ng build js-native --prod",
-    "gh": "npm run build-lib && ng build --prod --aot --ts-config=projects/demo/tsconfig.dist-lib.json --base-href=\"/ngx-modialog/\" && ./node_modules/.bin/angular-cli-ghpages --no-silent --dir dist/demo",
+    "build:lib": "ng-packagr -p projects/ngx-modialog/ng-package.json && cp README.md dist/ngx-modialog/",
+    "deploy:lib": "npm run build:lib && npm publish dist/ngx-modialog --registry https://registry.npmjs.org/",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
-    "@angular/http": "^6.0.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.0.0",
+    "@angular/animations": "^8.1.1",
+    "@angular/common": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@angular/forms": "^8.1.1",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.1.1",
     "core-js": "^2.5.4",
-    "rxjs": "^6.0.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.2.1",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.0",
-    "@angular-devkit/build-ng-packagr": "~0.6.0",
-    "@angular/cli": "~6.0.0",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/language-service": "^6.0.0",
+    "@angular-devkit/build-angular": "~0.801.1",
+    "@angular-devkit/build-ng-packagr": "~0.801.1",
+    "@angular/cli": "~8.1.1",
+    "@angular/compiler-cli": "^8.1.1",
+    "@angular/language-service": "^8.1.1",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -49,7 +48,7 @@
     "tsickle": ">=0.25.5",
     "tslib": "^1.7.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2",
+    "typescript": "3.4.5",
     "vex-js": "^4.1.0"
   }
 }
